style(portscan): split code from base.go
